refactor: changes types: double to bigdecimal

+++ src/main/java/com/audineia/sales_orders_service/dto/request/OrderItemRequestDTO.java
@@ -2,6 +2,8 @@ package com.audineia.sales_orders_service.dto.request;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Getter
 @Setter
@@ -10,7 +12,7 @@ import lombok.*;
 public class OrderItemRequestDTO {
     private Long productId;
     private Integer quantity;
-    private Double value;
+    private BigDecimal value;
 
     public Long getProductId() {
         return productId;
@@ -20,7 +22,7 @@ public class OrderItemRequestDTO {
         return quantity;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 }
